Ship the operator tools in the wheel; make an empty identity result honest
Round 4 review. No P1s and no data-loss findings this round -- both
reviewers independently re-derived and confirmed the default-flip
reasoning, the conditional-delete backends and the interaction surface. The
four P2s below are all real.

The operator tools now ship with the library, as the console commands
actingweb-verify-property-lists and actingweb-migrate-property-lists
(implementation moved to actingweb.maintenance; scripts/ keeps thin
wrappers). They were never in the wheel -- packages = [{include =
"actingweb"}] and no entry points -- which did not matter much until I made
converting existing lists an explicit operator step. Then the only
documented bulk path became a file a wheel install does not have. Verified
against a built wheel: modules present, both entry points registered.

verify(identity_key=...) now reports identity_checked_count, because an
empty duplicate_identities was ambiguous in the worst way. Rows lacking the
field are excluded from the comparison -- right, since bucketing them would
make each a duplicate of every other -- so "--identity-key id" against
items keyed on itemId excluded every row, compared nothing, and returned a
report shaped exactly like a clean one. That is this cycle's recurring
failure shape (a false-clean report poisoning an operator decision) with
the new check itself as the blind spot. Both tools now warn when the key
was supplied and nothing carried it.

Identity values are type-tagged before grouping. Python considers
True == 1 and hashes them identically, so {"id": true} and {"id": 1} were
grouped as one identity -- a duplicate that does not exist, enough to mark
a list unhealthy and fail a sweep.

--repair no longer compacts a v2 list flagged unhealthy only for duplicate
identities. compact() under v2 rebalances rank keys and nothing else, so
that rewrote every row -- including the documented interrupted-compaction
exposure -- and left the list just as unhealthy. The v2 report gained
needs_rebalance, which is the condition compact() can actually fix, and the
repair gate uses it instead of healthy.

Verified: ruff/pyright clean (bar the pre-existing migrate_db.py dotenv
error); 2753 passed / 26 skipped on DynamoDB, 827 passed / 18 skipped on
PostgreSQL; sphinx -W clean; wheel contents checked. The type-tag and
checked-count fixes are defeat-tested.
